// frontend/src/app/modules/openapi/model/updateTaskDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UpdateTaskDto { 
    title: string;
    description?: string;
    done?: boolean;
}